# Immutable object support without defensive copies