--- a/hazelcast/src/main/java/com/hazelcast/client/impl/protocol/task/cache/CacheGetAllMessageTask.java
+++ b/hazelcast/src/main/java/com/hazelcast/client/impl/protocol/task/cache/CacheGetAllMessageTask.java
@@ -71,7 +71,7 @@ public class CacheGetAllMessageTask
 
     @Override
     protected Object reduce(Map<Integer, Object> map) {
-        List<Map.Entry<Data, Data>> reducedMap = new ArrayList<>(map.size());
+        List<Map.Entry<Data, Object>> reducedMap = new ArrayList<>(map.size());
         for (Map.Entry<Integer, Object> entry : map.entrySet()) {
             MapEntries mapEntries = (MapEntries) nodeEngine.toObject(entry.getValue());
             mapEntries.putAllToList(reducedMap);