--- a/hazelcast/src/main/java/com/hazelcast/client/impl/protocol/task/map/MapValuesMessageTask.java
+++ b/hazelcast/src/main/java/com/hazelcast/client/impl/protocol/task/map/MapValuesMessageTask.java
@@ -45,7 +45,7 @@ public class MapValuesMessageTask
     protected Object reduce(Collection<QueryResultRow> result) {
         List<Data> values = new ArrayList<>(result.size());
         for (QueryResultRow resultEntry : result) {
-            values.add(resultEntry.getValue());
+            values.add(serializationService.toData(resultEntry.getValue()));
         }
         incrementOtherOperationsCount((MapService) getService(MapService.SERVICE_NAME), parameters);
         return values;