--- a/hazelcast/src/main/java/com/hazelcast/client/impl/protocol/task/replicatedmap/ReplicatedMapAddEntryListenerMessageTask.java
+++ b/hazelcast/src/main/java/com/hazelcast/client/impl/protocol/task/replicatedmap/ReplicatedMapAddEntryListenerMessageTask.java
@@ -48,10 +48,12 @@ public class ReplicatedMapAddEntryListenerMessageTask
     }
 
     @Override
-    protected ClientMessage encodeEvent(Data key, Data newValue, Data oldValue, Data mergingValue,
+    protected ClientMessage encodeEvent(Data key, Object newValue, Object oldValue, Object mergingValue,
                                         int type, UUID uuid, int numberOfAffectedEntries) {
-        return ReplicatedMapAddEntryListenerCodec.encodeEntryEvent(key, newValue,
-                oldValue, mergingValue, type, uuid, numberOfAffectedEntries);
+
+        return ReplicatedMapAddEntryListenerCodec.encodeEntryEvent(key, serializationService.toData(newValue),
+                serializationService.toData(oldValue), serializationService.toData(mergingValue), type, uuid,
+                numberOfAffectedEntries);
     }
 
     @Override