--- a/hazelcast/src/main/java/com/hazelcast/client/impl/protocol/task/replicatedmap/ReplicatedMapAddEntryListenerToKeyMessageTask.java
+++ b/hazelcast/src/main/java/com/hazelcast/client/impl/protocol/task/replicatedmap/ReplicatedMapAddEntryListenerToKeyMessageTask.java
@@ -49,10 +49,11 @@ public class ReplicatedMapAddEntryListenerToKeyMessageTask
     }
 
     @Override
-    protected ClientMessage encodeEvent(Data key, Data newValue, Data oldValue, Data mergingValue,
+    protected ClientMessage encodeEvent(Data key, Object newValue, Object oldValue, Object mergingValue,
                                         int type, UUID uuid, int numberOfAffectedEntries) {
-        return ReplicatedMapAddEntryListenerToKeyCodec.encodeEntryEvent(key, newValue,
-                oldValue, mergingValue, type, uuid, numberOfAffectedEntries);
+        return ReplicatedMapAddEntryListenerToKeyCodec.encodeEntryEvent(key, serializationService.toData(newValue),
+                serializationService.toData(oldValue), serializationService.toData(mergingValue), type, uuid,
+                numberOfAffectedEntries);
     }
 
     @Override