--- a/hazelcast/src/main/java/com/hazelcast/client/impl/protocol/task/replicatedmap/ReplicatedMapAddEntryListenerToKeyWithPredicateMessageTask.java
+++ b/hazelcast/src/main/java/com/hazelcast/client/impl/protocol/task/replicatedmap/ReplicatedMapAddEntryListenerToKeyWithPredicateMessageTask.java
@@ -50,10 +50,11 @@ public class ReplicatedMapAddEntryListenerToKeyWithPredicateMessageTask
     }
 
     @Override
-    protected ClientMessage encodeEvent(Data key, Data newValue, Data oldValue, Data mergingValue,
+    protected ClientMessage encodeEvent(Data key, Object newValue, Object oldValue, Object mergingValue,
                                         int type, UUID uuid, int numberOfAffectedEntries) {
-        return ReplicatedMapAddEntryListenerToKeyWithPredicateCodec.encodeEntryEvent(key, newValue,
-                oldValue, mergingValue, type, uuid, numberOfAffectedEntries);
+        return ReplicatedMapAddEntryListenerToKeyWithPredicateCodec.encodeEntryEvent(key,
+                serializationService.toData(newValue), serializationService.toData(oldValue),
+                serializationService.toData(mergingValue), type, uuid, numberOfAffectedEntries);
     }
 
     @Override