--- a/hazelcast/src/main/java/com/hazelcast/client/impl/protocol/task/replicatedmap/ReplicatedMapEntrySetMessageTask.java
+++ b/hazelcast/src/main/java/com/hazelcast/client/impl/protocol/task/replicatedmap/ReplicatedMapEntrySetMessageTask.java
@@ -21,6 +21,7 @@ import com.hazelcast.client.impl.protocol.codec.ReplicatedMapEntrySetCodec;
 import com.hazelcast.client.impl.protocol.task.AbstractPartitionMessageTask;
 import com.hazelcast.instance.impl.Node;
 import com.hazelcast.internal.nio.Connection;
+import com.hazelcast.internal.serialization.Data;
 import com.hazelcast.map.impl.MapEntries;
 import com.hazelcast.replicatedmap.impl.ReplicatedMapService;
 import com.hazelcast.replicatedmap.impl.operation.EntrySetOperation;
@@ -30,6 +31,7 @@ import com.hazelcast.security.permission.ReplicatedMapPermission;
 import com.hazelcast.spi.impl.operationservice.Operation;
 
 import java.security.Permission;
+import java.util.Map;
 
 public class ReplicatedMapEntrySetMessageTask
         extends AbstractPartitionMessageTask<String> {
@@ -51,7 +53,9 @@ public class ReplicatedMapEntrySetMessageTask
     @Override
     protected ClientMessage encodeResponse(Object response) {
         MapEntries entries = (MapEntries) response;
-        return ReplicatedMapEntrySetCodec.encodeResponse(entries.entries());
+        var collection = entries.entries().stream()
+                .map(e -> Map.entry(e.getKey(), serializationService.<Data>toData(e.getValue()))).toList();
+        return ReplicatedMapEntrySetCodec.encodeResponse(collection);
     }
 
     @Override