--- a/hazelcast/src/main/java/com/hazelcast/core/AggregatorCloneable.java
+++ b/hazelcast/src/main/java/com/hazelcast/core/AggregatorCloneable.java
@@ -1,0 +1,23 @@
+/*
+ * Copyright (c) 2008-2024, Hazelcast, Inc. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.hazelcast.core;
+
+import com.hazelcast.aggregation.Aggregator;
+
+public interface AggregatorCloneable<I, R> extends Aggregator<I, R> {
+    Aggregator cloneAggregator();
+}