… distribution/pom.xml
… distribution/src/assembly/assembly-descriptor-slim.xml
… distribution/src/assembly/assembly-descriptor.xml
… extensions/avro/pom.xml
… extensions/cdc-debezium/pom.xml
… extensions/cdc-mysql/pom.xml
… extensions/cdc-postgres/pom.xml
… extensions/csv/pom.xml
… extensions/elasticsearch/elasticsearch-7/pom.xml
… extensions/grpc/pom.xml
… extensions/hadoop-dist/files-azure/pom.xml
… extensions/hadoop-dist/files-gcs/pom.xml
… extensions/hadoop-dist/files-s3/pom.xml
… extensions/hadoop-dist/hadoop-all/pom.xml
… extensions/hadoop-dist/hadoop/pom.xml
… extensions/hadoop-dist/pom.xml
… extensions/hadoop/pom.xml
… extensions/kafka-connect/pom.xml
… extensions/kafka/pom.xml
… extensions/kinesis/pom.xml
… extensions/mapstore/pom.xml
… extensions/mongodb/pom.xml
… extensions/pom.xml
… extensions/protobuf/pom.xml
… extensions/python/pom.xml
… extensions/s3/pom.xml
… hazelcast-archunit-rules/pom.xml
… hazelcast-build-utils/pom.xml
… hazelcast-coverage-report/pom.xml
… hazelcast-it/distribution-it/pom.xml