This PR is more of a proof of concept. We are keen to see this feature added to Hazelcast.

Based this PR on the 5.5 release tag.



We make heavy use of Aggregations and `OnKey` operations.
We configure the maps to always store in `OBJECT` format for this.
The defensive copies cause considerable performance degradation when loading or replacing these objects.
We have zero replication on these objects, as all can be recreated. Take great care to only access them on the node, create a summary, and return it.

Second, when objects are moved between the nodes, they always get turned into a byte array, which is stored and then converted into the destination object. We also changed this to stream the Object to the other nodes, removing the need to allocate intermediate byte arrays. This effectively halved our memory load during rebalancing.


To demonstrate the performance difference, I wrote an image thumbnailed.
This thumbnailer loads a large image on one node. Then, using the execute on key can return a smaller image.
It also will store the smaller images locally.

code located here 
https://github.com/ashley-taylor/hazelcast-immutable-benchmark

before
```
Benchmark                 (nodes)   Mode  Cnt     Score     Error  Units
BenchmarkThumbnailer.get        1  thrpt   30  1685.872 ± 211.874  ops/s
BenchmarkThumbnailer.get        3  thrpt   30  1636.193 ± 190.123  ops/s
BenchmarkThumbnailer.get        5  thrpt   30  1703.851 ±  48.736  ops/s
BenchmarkThumbnailer.get        7  thrpt   30  1668.091 ±  99.055  ops/s
BenchmarkThumbnailer.get        9  thrpt   30  1635.359 ±  30.506  ops/s
```

after
```
Benchmark                 (nodes)   Mode  Cnt      Score      Error  Units
BenchmarkThumbnailer.get        1  thrpt   30  94386.273 ± 7238.175  ops/s
BenchmarkThumbnailer.get        3  thrpt   30  43732.168 ± 2008.540  ops/s
BenchmarkThumbnailer.get        5  thrpt   30  18613.176 ±  709.565  ops/s
BenchmarkThumbnailer.get        7  thrpt   30  19319.101 ±  911.382  ops/s
BenchmarkThumbnailer.get        9  thrpt   30  15770.779 ±  834.553  ops/s
```

Benchmark was run on a m2 MacBook pro.


Breaking changes (list specific methods/types/messages):
* API
* client protocol format
* serialized form
* snapshot format

Checklist:
- [ ] Labels (`Team:`, `Type:`, `Source:`, `Module:`) and Milestone set
- [ ] Add `Add to Release Notes` label if changes should be mentioned in release notes or `Not Release Notes content` if changes are not relevant for release notes
- [ ] Request reviewers if possible
- [ ] New public APIs have `@Nonnull/@Nullable` annotations
- [ ] New public APIs have `@since` tags in Javadoc
- [ ] Send backports/forwardports if fix needs to be applied to past/future releases