@devOpsHazelcast 2024-10-30
[![CLA assistant check](https://cla.hazelcast.org:443/pull/badge/not_signed)](https://cla.hazelcast.org:443/hazelcast/hazelcast?pullRequest=26453) <br/>Thank you for your submission! We really appreciate it. Like many open source projects, we ask that you all sign our [Contributor License Agreement](https://cla.hazelcast.org:443/hazelcast/hazelcast?pullRequest=26453) before we can accept your contribution.<br/>**6** out of **7** committers have signed the CLA.<br/><br/>:white_check_mark: frant-hartm<br/>:white_check_mark: JamesHazelcast<br/>:white_check_mark: burakgok<br/>:white_check_mark: oliverhowell<br/>:white_check_mark: devOpsHazelcast<br/>:white_check_mark: ashley-taylor<br/>:x: hz-devops-test<br/><sub>You have signed the CLA already but the status is still pending? Let us [recheck](https://cla.hazelcast.org:443/check/hazelcast/hazelcast?pullRequest=26453) it.</sub>